requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pytest==8.1.1
requests==2.31.0
typing_extensions==4.11.0
urllib3==2.2.1